refactor(SG): make the settings system lighter and easier
But yeah, Jesus Christ fixing the tests is going to take some time
